Split //common into /rpc and /exception
