+++ api/concept/type/AttributeType.java
@@ -21,7 +21,7 @@ package grakn.client.api.concept.type;
 
 import grakn.client.api.GraknTransaction;
 import grakn.client.api.concept.thing.Attribute;
-import grakn.client.common.GraknClientException;
+import grakn.client.common.exception.GraknClientException;
 import grakn.protocol.ConceptProto;
 
 import javax.annotation.CheckReturnValue;
@@ -29,7 +29,7 @@ import javax.annotation.Nullable;
 import java.time.LocalDateTime;
 import java.util.stream.Stream;
 
-import static grakn.client.common.ErrorMessage.Concept.BAD_VALUE_TYPE;
+import static grakn.client.common.exception.ErrorMessage.Concept.BAD_VALUE_TYPE;
 
 public interface AttributeType extends ThingType {
