+++ cluster/ClusterClient.java
@@ -22,8 +22,8 @@ package grakn.client.cluster;
 import grakn.client.api.GraknClient;
 import grakn.client.api.GraknOptions;
 import grakn.client.api.GraknSession;
-import grakn.client.common.GraknClientException;
-import grakn.client.common.GraknStub;
+import grakn.client.common.exception.GraknClientException;
+import grakn.client.common.rpc.GraknStub;
 import grakn.client.core.CoreClient;
 import grakn.common.collection.Pair;
 import grakn.protocol.ClusterServerProto;
@@ -36,8 +36,8 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
-import static grakn.client.common.ErrorMessage.Client.CLUSTER_UNABLE_TO_CONNECT;
-import static grakn.client.common.RequestBuilder.Cluster.Server.allReq;
+import static grakn.client.common.exception.ErrorMessage.Client.CLUSTER_UNABLE_TO_CONNECT;
+import static grakn.client.common.rpc.RequestBuilder.Cluster.Server.allReq;
 import static grakn.common.collection.Collections.pair;
 import static java.util.stream.Collectors.toMap;
 import static java.util.stream.Collectors.toSet;
