+++ concept/ConceptImpl.java
@@ -30,7 +30,7 @@ import grakn.client.api.concept.type.RelationType;
 import grakn.client.api.concept.type.RoleType;
 import grakn.client.api.concept.type.ThingType;
 import grakn.client.api.concept.type.Type;
-import grakn.client.common.GraknClientException;
+import grakn.client.common.exception.GraknClientException;
 import grakn.client.concept.thing.AttributeImpl;
 import grakn.client.concept.thing.EntityImpl;
 import grakn.client.concept.thing.RelationImpl;
@@ -43,7 +43,7 @@ import grakn.client.concept.type.ThingTypeImpl;
 import grakn.client.concept.type.TypeImpl;
 import grakn.protocol.ConceptProto;
 
-import static grakn.client.common.ErrorMessage.Concept.INVALID_CONCEPT_CASTING;
+import static grakn.client.common.exception.ErrorMessage.Concept.INVALID_CONCEPT_CASTING;
 import static grakn.common.util.Objects.className;
 
 public abstract class ConceptImpl implements Concept {
