+++ concept/thing/ThingImpl.java
@@ -24,7 +24,7 @@ import grakn.client.api.concept.thing.Attribute;
 import grakn.client.api.concept.thing.Thing;
 import grakn.client.api.concept.type.AttributeType;
 import grakn.client.api.concept.type.RoleType;
-import grakn.client.common.GraknClientException;
+import grakn.client.common.exception.GraknClientException;
 import grakn.client.concept.ConceptImpl;
 import grakn.client.concept.type.RoleTypeImpl;
 import grakn.client.concept.type.ThingTypeImpl;
@@ -34,17 +34,17 @@ import grakn.protocol.TransactionProto;
 import java.util.Objects;
 import java.util.stream.Stream;
 
-import static grakn.client.common.ErrorMessage.Concept.BAD_ENCODING;
-import static grakn.client.common.ErrorMessage.Concept.MISSING_IID;
-import static grakn.client.common.ErrorMessage.Concept.MISSING_TRANSACTION;
-import static grakn.client.common.RequestBuilder.Thing.deleteReq;
-import static grakn.client.common.RequestBuilder.Thing.getHasReq;
-import static grakn.client.common.RequestBuilder.Thing.getPlayingReq;
-import static grakn.client.common.RequestBuilder.Thing.getRelationsReq;
-import static grakn.client.common.RequestBuilder.Thing.isInferredReq;
-import static grakn.client.common.RequestBuilder.Thing.protoThing;
-import static grakn.client.common.RequestBuilder.Thing.setHasReq;
-import static grakn.client.common.RequestBuilder.Thing.unsetHasReq;
+import static grakn.client.common.exception.ErrorMessage.Concept.BAD_ENCODING;
+import static grakn.client.common.exception.ErrorMessage.Concept.MISSING_IID;
+import static grakn.client.common.exception.ErrorMessage.Concept.MISSING_TRANSACTION;
+import static grakn.client.common.rpc.RequestBuilder.Thing.deleteReq;
+import static grakn.client.common.rpc.RequestBuilder.Thing.getHasReq;
+import static grakn.client.common.rpc.RequestBuilder.Thing.getPlayingReq;
+import static grakn.client.common.rpc.RequestBuilder.Thing.getRelationsReq;
+import static grakn.client.common.rpc.RequestBuilder.Thing.isInferredReq;
+import static grakn.client.common.rpc.RequestBuilder.Thing.protoThing;
+import static grakn.client.common.rpc.RequestBuilder.Thing.setHasReq;
+import static grakn.client.common.rpc.RequestBuilder.Thing.unsetHasReq;
 import static grakn.client.concept.type.TypeImpl.protoTypes;
 import static grakn.common.util.Objects.className;
 import static java.util.Arrays.asList;
