+++ concept/type/EntityTypeImpl.java
@@ -28,7 +28,7 @@ import grakn.protocol.ConceptProto;
 
 import java.util.stream.Stream;
 
-import static grakn.client.common.RequestBuilder.Type.EntityType.createReq;
+import static grakn.client.common.rpc.RequestBuilder.Type.EntityType.createReq;
 
 public class EntityTypeImpl extends ThingTypeImpl implements EntityType {
