+++ concept/type/TypeImpl.java
@@ -30,7 +30,7 @@ import grakn.client.api.concept.type.RelationType;
 import grakn.client.api.concept.type.RoleType;
 import grakn.client.api.concept.type.ThingType;
 import grakn.client.api.concept.type.Type;
-import grakn.client.common.GraknClientException;
+import grakn.client.common.exception.GraknClientException;
 import grakn.client.common.Label;
 import grakn.client.concept.ConceptImpl;
 import grakn.client.concept.thing.AttributeImpl;
@@ -46,16 +46,16 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Stream;
 
-import static grakn.client.common.ErrorMessage.Concept.BAD_ENCODING;
-import static grakn.client.common.ErrorMessage.Concept.INVALID_CONCEPT_CASTING;
-import static grakn.client.common.ErrorMessage.Concept.MISSING_LABEL;
-import static grakn.client.common.ErrorMessage.Concept.MISSING_TRANSACTION;
-import static grakn.client.common.RequestBuilder.Type.deleteReq;
-import static grakn.client.common.RequestBuilder.Type.getSubtypesReq;
-import static grakn.client.common.RequestBuilder.Type.getSupertypeReq;
-import static grakn.client.common.RequestBuilder.Type.getSupertypesReq;
-import static grakn.client.common.RequestBuilder.Type.isAbstractReq;
-import static grakn.client.common.RequestBuilder.Type.setLabelReq;
+import static grakn.client.common.exception.ErrorMessage.Concept.BAD_ENCODING;
+import static grakn.client.common.exception.ErrorMessage.Concept.INVALID_CONCEPT_CASTING;
+import static grakn.client.common.exception.ErrorMessage.Concept.MISSING_LABEL;
+import static grakn.client.common.exception.ErrorMessage.Concept.MISSING_TRANSACTION;
+import static grakn.client.common.rpc.RequestBuilder.Type.deleteReq;
+import static grakn.client.common.rpc.RequestBuilder.Type.getSubtypesReq;
+import static grakn.client.common.rpc.RequestBuilder.Type.getSupertypeReq;
+import static grakn.client.common.rpc.RequestBuilder.Type.getSupertypesReq;
+import static grakn.client.common.rpc.RequestBuilder.Type.isAbstractReq;
+import static grakn.client.common.rpc.RequestBuilder.Type.setLabelReq;
 import static grakn.client.concept.type.RoleTypeImpl.protoRoleTypes;
 import static grakn.client.concept.type.ThingTypeImpl.protoThingType;
 import static grakn.common.util.Objects.className;
