+++ core/CoreDatabaseManager.java
@@ -21,16 +21,16 @@ package grakn.client.core;
 
 import grakn.client.api.database.Database;
 import grakn.client.api.database.DatabaseManager;
-import grakn.client.common.GraknClientException;
-import grakn.client.common.GraknStub;
+import grakn.client.common.exception.GraknClientException;
+import grakn.client.common.rpc.GraknStub;
 
 import java.util.List;
 
-import static grakn.client.common.ErrorMessage.Client.DB_DOES_NOT_EXIST;
-import static grakn.client.common.ErrorMessage.Client.MISSING_DB_NAME;
-import static grakn.client.common.RequestBuilder.Core.DatabaseManager.allReq;
-import static grakn.client.common.RequestBuilder.Core.DatabaseManager.containsReq;
-import static grakn.client.common.RequestBuilder.Core.DatabaseManager.createReq;
+import static grakn.client.common.exception.ErrorMessage.Client.DB_DOES_NOT_EXIST;
+import static grakn.client.common.exception.ErrorMessage.Client.MISSING_DB_NAME;
+import static grakn.client.common.rpc.RequestBuilder.Core.DatabaseManager.allReq;
+import static grakn.client.common.rpc.RequestBuilder.Core.DatabaseManager.containsReq;
+import static grakn.client.common.rpc.RequestBuilder.Core.DatabaseManager.createReq;
 import static java.util.stream.Collectors.toList;
 
 public class CoreDatabaseManager implements DatabaseManager {
