+++ stream/BidirectionalStream.java
@@ -19,8 +19,8 @@
 
 package grakn.client.stream;
 
-import grakn.client.common.GraknClientException;
-import grakn.client.common.GraknStub;
+import grakn.client.common.exception.GraknClientException;
+import grakn.client.common.rpc.GraknStub;
 import grakn.protocol.TransactionProto.Transaction.Req;
 import grakn.protocol.TransactionProto.Transaction.Res;
 import grakn.protocol.TransactionProto.Transaction.ResPart;
@@ -34,8 +34,8 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
-import static grakn.client.common.ErrorMessage.Client.UNKNOWN_REQUEST_ID;
-import static grakn.client.common.ErrorMessage.Internal.ILLEGAL_ARGUMENT;
+import static grakn.client.common.exception.ErrorMessage.Client.UNKNOWN_REQUEST_ID;
+import static grakn.client.common.exception.ErrorMessage.Internal.ILLEGAL_ARGUMENT;
 import static java.util.Spliterator.IMMUTABLE;
 import static java.util.Spliterator.ORDERED;
 import static java.util.Spliterators.spliteratorUnknownSize;
